feat(ci): dedicated multi-device test job + eca.py cleanup
Follow-up to #961. Add .github/workflows/test_multidevice.yml — a standalone
CI job that sets XLA_FLAGS in the shell environment before pytest starts, so
tests in tests/test_multidevice/ always see 2 CPU devices reliably. The
existing test job is untouched.

Move the ECA multi-device regression test to tests/test_multidevice/test_eca.py.
Removes the fragile module-level os.environ trick (unreliable under pytest-xdist)
and adds a NamedSharding assertion that makes the guarded failure mode explicit.

Clean up eca.py post-#961:
- Trim the 8-line comment to one line
- Fix latent crash: superchain_size=None fell into the else branch and raised
  TypeError on num_chains // None; guard with `is None or == 1`
